(PE-11531) Use clientversion fact on service mgmt gating
Prior to this commit we were determining whether to manage the
service or not based on the aio_agent_version fact and the is_pe
variable. The is_pe check is redundant because if the module
is being run on a non-PE node, the !is_pe && client_version > 4.0.0
check earlier in the `puppet_agent` class would have already
stopped compilation.
This commit updates the service management to only occur if the
client_version is < 4.0.0. We must do this because PE already
manages the AIO services as part of the Agent nodegroup, so if
this class also managed those resources, we would see conflicts.