(PE-11531) Do not test pre-4 file/ssl/mco migration on post-4
Prior to this commit the specs were not updated to match the
behavior of post-4 upgrade. Specifically, they were still attempting
to test that the file/mco/ssl migration resources were being created
but they are not needed on a post-4 upgrade.
This commit removes the unnecessary resource tests and replaces them
with tests to ensure the migration classes are not included on post-4
upgrades.