(PE-11531) Test solaris, fix package resource for Sol10
Prior to this commit we were not running specs for Solaris.
This commit enables the specs, and adds the `package_version`
parameter for 4+ puppet. When adding the specs, I noticed that
the solaris 10 package provider doesn't allow `verionable`, we
we'll just have to pass in `present` and hope that the right
package version was setup elsewhere.