(PE-11531) Always set _package_version
Prior to this commit `_package_version` would only be
set if it was set and `package_version` differed from
`::clientversion`. This isn't really necessary since
the default for `_package_version` is `undef`, so if
we just check undef elsewhere, we'll know whether a
user wants a specific version or not.
This commit changes the `_package_version` to always
be set to `package_version` (which can be `undef`).