(PE-11531) Allow specifying package version for Windows
Prior to this commit we assumed that the package filename
on windows would not have a version number. This was only
true for PE builds. FOSS builds do have the package version
in the filename.
This commit adds the `_package_version` to the Windows MSI
file name, if it is specified and if we are not a PE install.