(PE-11531) Stop puppet 4+ upgrades if no agent_version is specified
Prior to this commit we had completely removed the restriction on
upgrading Puppet 4+ agents. This would cause the module to attempt
to upgrade agents even if the `$package_version` param had not been
set. This would result in attempting to install non-existent packages
missing the package version and other problems. Unless we want to just
always default to the master agent version, I suggest that we only
upgrade if a `$package_version` is specified.
This commit changes the module to only upgrade Puppet 4+ agents if
a `$package_version` has been specified, as well as making the usage
of `$_package_version` more consistent after `$package_version` has
been assigned to it.