(PE-11531) Require package_version
Prior to this commit we did not actually require that `package_version`
was set, even when we did not have any way of reasonably guessing a
version, such as in FOSS. Additionally, after upgrading to 4+, a user
most likely will not want to immediately want to upgrade to whatever
version of Puppet their Master is running.
This commit makes `package_version` required if you aren't running
PE, or you are running PE and your agent version is already 4+.
Additionally, many of the specs had to be re-worked to support this
new required parameter.