(PE-11531) Use _package_version for aix
Prior to this commit we were for some reason always using the
`master_agent_version` for aix packages. While this is a sensible
default, a user may want to use a different version so we should
make use of the `_package_version`
This commit updates the aix package filename to use the
`_package_version`.