(PE-11531) Ensure package 'present' on non-versionable platforms
Prior to this commit the enforcement of `ensure => present` on platforms
that require `present` because they do not support `versionable` was
inconsistent.
This commit should specifically catch all the platforms that do not
support versionable. More testing will need to be done to ensure that
this is really the case.