(PE-11531) Fixup windows specs, correct PE Win guard
Prior to this commit the windows specs would not run on
anything other than 3.7/4.x, so that restriction has
been removed. Additionally, windows needed a spec to test
the upgrading to the master's agent version by default behavior,
so one was added.

In the process of adding that, we discovered
that the windows guard would, in fact, never guard anything
because it was checking a non-existent `agent_version` fact
when the correct fact would have been `aio_agent_version`, which
has now been added.

Also corrects the failed rebase some more by removing the
3.8 guard for init.pp